[WIP] Create StreamServer class
